feat:  Add setting/getting properties on directories. (#12247)
Summary:

We'd like to set/get properties on directories if the remote filesystem supports APIs like TTL. In POSIX, this would be similar to ioctl.

Simply expose a property bag for which users can use for setting key-value pairs for the set path, and on the get path, only support getting a single key for now. If the key is not present, nullopt is returned.

Reviewed By: singcha

Differential Revision: D68650034